+++ R/create.R
@@ -32,7 +32,7 @@ NULL
 #' creation section and examples)_.
 #' * Via technology specific convenience functions _(exported from \{Giotto\})_
 #'
-#' See \url{https://drieslab.github.io/Giotto_website/articles/object_creation.html}
+#' See \url{https://giotto-suite.github.io/Giotto_website/articles/object_creation.html}
 #' for more information. The details sections here also expand on the above
 #' options.
 #' @param gpolygons giotto polygons
@@ -167,7 +167,7 @@ NULL
 #' Giotto::createGiottoCosMxObject()
 #' }
 #'
-#' Also see \url{https://drieslab.github.io/Giotto_website/articles/import_utilities.html}
+#' Also see \url{https://giotto-suite.github.io/Giotto_website/articles/import_utilities.html}
 #' for technology-specific import utilities that can be used with the piecewise
 #' `giotto` object assembly method.
 #'
@@ -1289,20 +1289,39 @@ createGiottoObjectSubcellular <- function(
 
 #' @title Create S4 exprObj
 #' @name createExprObj
-#' @description Create an S4 exprObj
+#' @description Create an `exprObj` (thin metadata wrapper/adapter) for
+#'   Giotto expression value information. `matrix` and tabular formats
+#'   will be coerced to {Matrix} sparse format `dgCMatrix`. For tabular
+#'   inputs, the first column is used as row names.
+#'
+#' ## Alternative and disk-backed formats
+#'
+#' Pre-constructed backed matrices (`HDF5Array`, `dbMatrix`, `tiledb_array`,
+#' `IterableMatrix`) may also be passed directly to `expression_data` and will
+#' be used as-is.
 #' @inheritParams data_access_params
-#' @param expression_data expression data
+#' @param expression_data expression data. Accepts matrix-like data and
+#'   coercible `data.frame` formats (first column used as row names). A
+#'   filepath to a flat file readable by [data.table::fread()] can also be
+#'   passed.
 #' @param name name of exprObj
 #' @param provenance origin data of expression information (if applicable)
-#' @param misc misc
-#' @param expression_matrix_class class of expression matrix to
-#' use (e.g. 'dgCMatrix', 'DelayedArray', 'dbMatrix')
-#' @returns exprObj
+#' @param misc misc about the object to attach
+#' @param expression_matrix_class deprecated. Previously wrapped the matrix in
+#'   a `DelayedArray` when selected. Moving forward, pass a pre-constructed 
+#'   backed matrix to `expression_data` directly if needed.
+#' @returns `exprObj` wrapping the expression matrix with `spat_unit`,
+#'   `feat_type`, and provenance metadata attached.
 #' @examples
+#' m <- matrix(rpois(100, 1), nrow = 10,
+#'     dimnames = list(paste0("feat", 1:10), paste0("cell", 1:10))
+#' )
+#' createExprObj(m)
+#'
+#' # example 0-dominated sparse matrix
 #' x_expr <- readRDS(system.file("extdata/toy_matrix.RDS",
 #'     package = "GiottoClass"
 #' ))
-#'
 #' createExprObj(expression_data = x_expr)
 #' @export
 createExprObj <- function(
@@ -1312,9 +1331,15 @@ createExprObj <- function(
         feat_type = "rna",
         provenance = NULL,
         misc = NULL,
-        expression_matrix_class = c("dgCMatrix", "DelayedArray", "dbMatrix")) {
+        expression_matrix_class = deprecated()) {
+    if (is_present(expression_matrix_class)) {
+        warning(sprintf(
+            "[createExprObj] param '%s' is deprecated", 
+            "expression_matrix_class"), 
+        call. = FALSE)
+    }
+  
     exprMat <- .evaluate_expr_matrix(expression_data,
-        expression_matrix_class = expression_matrix_class,
         feat_type = feat_type
     )
 
@@ -1328,30 +1353,12 @@ createExprObj <- function(
     )
 }
 
-
-#' @title create_expr_obj
-#' @name create_expr_obj
-#' @param exprMat matrix of expression information
-#' @keywords internal
-#' @returns expr_obj
-#' @examples
-#' x_expr <- readRDS(system.file("extdata/toy_matrix.RDS",
-#'     package = "GiottoClass"
-#' ))
-#'
-#' create_expr_obj(exprMat = x_expr)
-#'
-#' @export
 create_expr_obj <- function(name = "test",
     exprMat = NULL,
     spat_unit = "cell",
     feat_type = "rna",
     provenance = NULL,
     misc = NULL) {
-    deprecate_soft("3.3.0",
-        what = "create_expr_obj()",
-        with = "createExprObj()"
-    )
 
     if (is.null(exprMat)) exprMat <- matrix()
 
@@ -1950,7 +1957,7 @@ createSpatEnrObj <- function(enrichment_data,
     create_spat_enr_obj(
         name = name,
         method = method,
-        enrichDT = enrichment_data,
+        enrichDT = enrichDT,
         spat_unit = spat_unit,
         feat_type = feat_type,
         provenance = provenance,
@@ -3031,8 +3038,11 @@ create_giotto_polygon_object <- function(name = "cell",
 
 #' @title createGiottoImage
 #' @name createGiottoImage
-#' @description Creates a giotto image that can be added to a Giotto object
-#' and/or used to add an image to the spatial plotting functions
+#' @description **SUPERSEDED** *[createGiottoLargeImage()] is preferred for most
+#' uses*. This function may be renamed and/or replaced in the future.
+#'
+#' Creates a [giottoImage] that can be added to a Giotto object and/or used to
+#' add an image to the spatial plotting functions.
 #' @inheritParams data_access_params
 #' @param spatial_locs spatial locations (alternative if \code{gobject = NULL})
 #' @param spat_loc_name name of spatial locations within gobject
@@ -3136,7 +3146,7 @@ createGiottoImage <- function(gobject = NULL,
             if (transf == "flip_x_axis") {
                 mg_object <- magick::image_flop(mg_object)
             } else if (transf == "flip_y_axis") {
-                mg_object <- magick::image_flop(mg_object)
+                mg_object <- magick::image_flip(mg_object)
             } else {
                 wrap_msg(transf, " is not a supported transformation, see
                     details")